common submodule added and moved common functionalities to it, created an npm module using it
