+++ auth/package.json
@@ -7,7 +7,7 @@
     "start": "ts-node-dev --poll src/index.ts",
     "test": "jest --watchAll --no-cache"
   },
-  "jest":{
+  "jest": {
     "preset": "ts-jest",
     "testEnvironment": "node",
     "setupFilesAfterEnv": [
@@ -17,6 +17,7 @@
   "author": "abin",
   "license": "ISC",
   "dependencies": {
+    "@abitickets/common": "^1.0.5",
     "@types/cookie-session": "^2.0.45",
     "@types/express": "^4.17.17",
     "@types/jsonwebtoken": "^9.0.2",
