--- auth/src/errors/bad-request-error.ts
@@ -1,14 +1,0 @@
-import { CustomError } from "./custom-error";
-
-export class BadRequestError extends CustomError {
-	statusCode = 400;
-
-	constructor(public message: string) {
-		super(message);
-		Object.setPrototypeOf(this, BadRequestError.prototype);
-	}
-
-	serializeErrors() {
-		return [{ message: this.message }];
-	}
-}
