--- auth/src/errors/custom-error.ts
@@ -1,11 +1,0 @@
-export abstract class CustomError extends Error {
-	abstract statusCode: number;
-
-	constructor(message: string) {
-		// this message is only for logging purpose,not to send to user
-		super(message);
-		Object.setPrototypeOf(this, CustomError.prototype);
-	}
-
-	abstract serializeErrors(): { message: string; field?: string }[];
-}
