--- auth/src/errors/database-connection-error.ts
@@ -1,15 +1,0 @@
-import { CustomError } from "./custom-error";
-
-export class DatabaseConnectionError extends CustomError {
-	statusCode = 500;
-	reason = "error connecting to database";
-	constructor() {
-		super('error connecting to db');
-
-		Object.setPrototypeOf(this, DatabaseConnectionError.prototype);
-	}
-
-	serializeErrors() {
-		return [{ message: this.reason }];
-	}
-}
