--- auth/src/errors/not-authorized-error.ts
@@ -1,15 +1,0 @@
-import { CustomError } from "./custom-error";
-
-export class NotAuthorizedError extends CustomError {
-	statusCode = 401;
-
-	constructor() {
-		super("Not Authorized");
-
-		Object.setPrototypeOf(this, NotAuthorizedError.prototype);
-	}
-
-	serializeErrors() {
-		return [{ message: "Not Authorized" }];
-	}
-}
