--- auth/src/errors/not-found-error.ts
@@ -1,14 +1,0 @@
-import { CustomError } from "./custom-error";
-
-export class NotFoundError extends CustomError {
-	statusCode = 404;
-
-	constructor() {
-		super("Route not found");
-		Object.setPrototypeOf(this, NotFoundError.prototype);
-	}
-
-	serializeErrors() {
-		return [{ message: "Not found" }];
-	}
-}
