--- auth/src/errors/request-validation-error.ts
@@ -1,21 +1,0 @@
-import { ValidationError } from "express-validator";
-import { CustomError } from "./custom-error";
-
-export class RequestValidationError extends CustomError {
-	statusCode = 400;
-	constructor(public errors: ValidationError[]) {
-		super('invalid request parameters');
-
-		// only because we are extending a built in class
-		Object.setPrototypeOf(this, RequestValidationError.prototype);
-	}
-
-	serializeErrors() {
-		return this.errors.map((err) => {
-			if (err.type === "field") {
-				return { message: err.msg, field: err.path };
-			}
-			return { message: err.msg };
-		});
-	}
-}
