+++ auth/src/routes/signin.ts
@@ -2,9 +2,8 @@ import express, { Request, Response } from "express";
 import { body } from "express-validator";
 import jwt from "jsonwebtoken";
 
-import { PasswordManage } from "../../services/password";
-import { validateRequest } from "../middlewares/validate-request";
-import { BadRequestError } from "../errors/bad-request-error";
+import { validateRequest, BadRequestError } from "@abitickets/common";
+import { PasswordManage } from "../services/password";
 import { User } from "../models/user";
 
 const router = express.Router();
